// types/schemas.ts
export const statusValidationSchema: ZodType<StatusFormData> = z.object({
  status: z.enum([
    "pending",
    "processing",
    "onhold",
    "completed",
    "refunded",
    "failed",
  ]),
});