// types/models.ts
export type statusStoreEnum = "online" | "pending" | "suspended";
export type statusAddress = "draft" | "archive";
export type statusCartEnum = "draft" | "abandoned" | "completed";
export type badge = "hot" | "bestdeals" | "sale" | "off" | "soldout" | "none";
export type Image = {
  url: string;
};

export type TypeNewsletterModel = {
  email: string;
};

export type TypeTokenModel = {
  template: string;
  token: string;
  status: "draft" | "publish" | "archive";
  user_id: string;
};

export type TypeDiscountModel = {
  _id: string;
  code: string;
  slug: string;
  startDate: Date;
  endDate: Date;
  validity: "once";
  discount: number;
  status: "available" | "expired";
};

export type TypeAddressModel = {
  firstName: string;
  lastName: string;
  address: string;
  city: string;
  state: string;
  phone: string;
  country?: string;
  zipCode: string;
  email: string;
  user_id: string;
  status: statusAddress;
};

export type TypeCartItemModel = {
  cart: TypeCartModel;
  variant: TypeProductVariantModel;
  productName: string;
  productImage: string;
  store: TypeStoreModel;
  qty: number;
};

export type TypeCartModel = {
  cartItems: TypeCartItemModel[];
  subTotal: number;
  coupon: number;
  shipping: TypeShippingModel;
  user_id: string;
  status: statusCartEnum;
};

export type TypeStoreModel = {
  _id: string;
  user_id: string; //clerk user_id
  name: string;
  description: string;
  logo: string;
  products: TypeProductModel[];
  orderitems: TypeOrderItemModel[];
  slides: TypeSlideModel[];
  subscription: TypeSubscriptionModel;
  status: statusStoreEnum;
  createdAt: Date;
};

export type TypeTrackOrderModel = {
  _id: string;
  orderitem: TypeOrderModel;
  trackactivity: TypeTrackActivityModel[];
  status: "open" | "packaging" | "onroad" | "delivered" | "failed";
  user_id: string;
};

export type TypeTrackActivityModel = {
  _id: string;
  trackorder: TypeTrackOrderModel;
  activity: string;
  user_id: string;
};

export type TypeImageModel = {
  url: string;
  store: TypeStoreModel;
  user_id: string;
};

export type TypeProductModel = {
  _id: string;
  featured?: boolean;
  badge?: badge;
  name: string;
  slug: string;
  description: string;
  additionnal: string;
  specification: string;
  store: TypeStoreModel;
  collections: TypeCollectionModel;
  tags: TypeCollectionModel;
  category: TypeCategoryModel;
  subCategories: TypeSubCategoryModel[];
  brand: TypeBrandModel;
  details: TypeDetailModel[];
  questions: TypeQuestionModel[];
  reviews: TypeReviewModel[];
  productVariants: TypeProductVariantModel[];
  images: string[];
  price: number;
  discount: number;
  seoSlug?: string;
  seoDescription?: string;
  seoTitle?: string;
  status: "draft" | "publish" | "archive";
  inventory: "instock" | "outstock";
  weight: number;
  sku: string;
  unit: string;
  user_id: string;
  views: number;
  createdAt?: Date;
};

export type TypeProductVariantModel = {
  productId: TypeProductModel;
  name: string;
  color: TypeColorModel;
  colorImages: Image[];
  size?: TypeSizeModel;
  sizeImages?: Image[];
  weight: number;
  inventory: "instock" | "outstock";
  sku: string;
  price: number;
  discount: number;
  colorValue?: string;
  status: "draft" | "publish" | "archive";
};
// ...
export type TypeShippingModel = {
  _id: string;
  name: string;
  description: string;
  slug: string;
  image: string;
  user_id: string;
  status: "draft" | "publish" | "archive";
  delay: number;
  fixed_amount: number;
  fees: number;
  region: string[];
  unit_price_weight: number;
  price_range_start: number;
  price_range_end: number;
  createdAt: Date;
  store: TypeStoreModel;
};

// export type TypeRegionModels = {
//   image: string;
//   name: string;
// };

export type TypePmethodModel = {
  _id: string;
  name: string;
  description: string;
  slug: string;
  image: string;
  user_id: string;
  status: "draft" | "publish" | "archive";
  createdAt: Date;
};
// ...
export type TypeSlideModel = {
  _id: string;
  name: string;
  description: string;
  slug: string;
  image: string;
  user_id: string;
  status: "draft" | "publish" | "archive";
  slideItem: string[];
  createdAt: Date;
};
// ...
export type TypeOrderModel = {
  _id: string;
  subTotal: number;
  shipping: number;
  discount: number;
  tax: number;
  total: number;
  pmethod: TypePmethodModel;
  discountItem: TypeDiscountModel;
  delivered: boolean;
  earning: number;
  status:
    | "pending"
    | "processing"
    | "onhold"
    | "completed"
    | "refunded"
    | "failed";
  orderitems: TypeOrderItemModel[];
  customer: string;
};
export type TypeOrderItemModel = {
  _id: string;
  status:
    | "pending"
    | "processing"
    | "onhold"
    | "completed"
    | "refunded"
    | "failed";
  cartItem: TypeCartItemModel;
  order: TypeOrderModel;
  shipping: TypeShippingModel;
  address: TypeAddressModel;
  store: TypeStoreModel;
  shippingAmount: number;
  delivered: boolean;
  earning: number;
  paidAmount: number;
  trackorder: TypeTrackOrderModel;
  createdAt: Date;
};
export type TypeOptionsModel = { _id: string };
export type TypeStyleModel = {
  _id: string;
  color: string;
  image: string;
};

export type TypeSubscriptionModel = {
  store: TypeStoreModel;
  payments: string[];
  startDate: Date;
  endDate?: Date | undefined;
  status: "active" | "ended" | "cancelled";
  type: "free" | "pro";
  user_id: string;
};